add ability to edit duration of each move type
