+++ src/css/components/EditValues.css
@@ -3,11 +3,11 @@
 	width: 100%;
 }
 
-.EditProbsButton {
+.EditValuesButton {
 	width: 100%;
 }
 
-.EditProbsContainer {
+.EditValuesContainer {
 	display: inline-block;
 	text-align: center;
 	width: 38%;
