+++ src/css/containers/TrainingView.css
@@ -4,12 +4,6 @@
 	width: 15%;
 }
 
-.TrainingButton {
-	margin-top: 3%;
-	margin-right: 3%;
-	width: 45%;
-}
-
 .SlidingContainer {
 	height: 15vh;
 	border: 1px solid #d9d9d9;
@@ -21,4 +15,9 @@
     overflow-x: auto;
     overflow: hidden;
     padding: 0.75%;
+}
+
+.ButtonsDiv {
+	width: 100%;
+	display: inline-block;
 }
